// apps/web-svelte/src/lib/api/generated/model/externalLibraryMount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ExternalLibraryMount {
  id: string;
  connectionId: string;
  libraryRootId: string;
  remoteRootId: string;
  remotePath: string;
  localPath: string;
  label: string;
}